+++ lib/presently/state.rb
@@ -3,6 +3,7 @@
 # Released under the MIT License.
 # Copyright, 2026, by Samuel Williams.
 
+require "console"
 require "json"
 
 module Presently